Fix includes after move
